comment.py 내용 post.py에 이관
